+++ app/src/main/java/com/mobven/fitai/login/sign_in/AuthFragment.kt
@@ -1,4 +1,4 @@
-package com.mobven.fitai.login.sign_up
+package com.mobven.fitai.login.sign_in
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,19 @@ import android.view.View
 import android.view.ViewGroup
 import com.mobven.fitai.R
 
-class SignUpFragment : Fragment() {
+class AuthFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_sign_up, container, false)
+
+        return inflater.inflate(R.layout.fragment_auth, container, false)
     }
 
 }
