+++ app/src/main/java/com/mobven/fitai/login/sign_up/screens/DailyMovementFragment.kt
@@ -1,4 +1,4 @@
-package com.mobven.fitai.login.sign_up
+package com.mobven.fitai.login.sign_up.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,58 +6,58 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.mobven.fitai.R
-import com.mobven.fitai.databinding.FragmentDietTypeBinding
+import com.mobven.fitai.databinding.FragmentDailyMovementBinding
 import com.mobven.fitai.login.sign_up.adapter.SignUpListAdapter
 import com.mobven.fitai.login.sign_up.model.SignUpSelectorItem
 import com.mobven.fitai.util.SignUpSelectorType
 
-class DietTypeFragment : Fragment() {
 
-    private lateinit var binding: FragmentDietTypeBinding
-    private val adapter = SignUpListAdapter()
+class DailyMovementFragment : Fragment() {
+
+    private lateinit var binding: FragmentDailyMovementBinding
+    private var adapter = SignUpListAdapter()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentDietTypeBinding.inflate(inflater, container, false)
+        binding = FragmentDailyMovementBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val dietTypeList = listOf(
+        val dailyMovementList = listOf(
             SignUpSelectorItem(
-                title = getString(R.string.ketogenic),
+                title = getString(R.string.almost_never),
                 isSelected = false,
-                type = SignUpSelectorType.RADIO
+                type = SignUpSelectorType.RADIO,
             ),
             SignUpSelectorItem(
-                title = getString(R.string.pescetarian),
+                title = getString(R.string.rarely),
                 isSelected = false,
-                type = SignUpSelectorType.RADIO
+                type = SignUpSelectorType.RADIO,
             ),
             SignUpSelectorItem(
-                title = getString(R.string.vegan),
+                title = getString(R.string.sometimes),
                 isSelected = false,
-                type = SignUpSelectorType.RADIO
+                type = SignUpSelectorType.RADIO,
             ),
             SignUpSelectorItem(
-                title = getString(R.string.vegetarian),
+                title = getString(R.string.often),
                 isSelected = false,
-                type = SignUpSelectorType.RADIO
+                type = SignUpSelectorType.RADIO,
             ),
             SignUpSelectorItem(
-                title = getString(R.string.traditional),
+                title = getString(R.string.always),
                 isSelected = false,
-                type = SignUpSelectorType.RADIO
+                type = SignUpSelectorType.RADIO,
             )
-        )
+        ).reversed()
 
-        adapter.submitList(dietTypeList)
-        binding.rvDietType.adapter = adapter
+        adapter.submitList(dailyMovementList)
+        binding.rvDailyMovement.adapter = adapter
 
     }
-
 }
