+++ app/src/main/java/com/mobven/fitai/login/sign_up/screens/GenderSignUpFragment.kt
@@ -1,4 +1,4 @@
-package com.mobven.fitai.login.sign_up
+package com.mobven.fitai.login.sign_up.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,51 +6,47 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.mobven.fitai.R
-import com.mobven.fitai.databinding.FragmentGoalsBinding
+import com.mobven.fitai.databinding.FragmentGenderSignUpBinding
 import com.mobven.fitai.login.sign_up.adapter.SignUpListAdapter
 import com.mobven.fitai.login.sign_up.model.SignUpSelectorItem
 import com.mobven.fitai.util.SignUpSelectorType
 
-class GoalsFragment : Fragment() {
+class GenderSignUpFragment : Fragment() {
 
-    private lateinit var binding: FragmentGoalsBinding
+    private lateinit var binding: FragmentGenderSignUpBinding
     private val adapter = SignUpListAdapter()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentGoalsBinding.inflate(inflater, container, false)
+        binding = FragmentGenderSignUpBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val goalsList = listOf(
+        val genderList = listOf(
             SignUpSelectorItem(
-                title = getString(R.string.calorie_control),
+                title = getString(R.string.other_gender),
                 isSelected = false,
                 type = SignUpSelectorType.RADIO
             ),
             SignUpSelectorItem(
-                title = getString(R.string.muscle_gain),
+                title = getString(R.string.male_gender),
                 isSelected = false,
                 type = SignUpSelectorType.RADIO
             ),
             SignUpSelectorItem(
-                title = getString(R.string.weight_gain),
-                isSelected = false,
-                type = SignUpSelectorType.RADIO
-            ),
-            SignUpSelectorItem(
-                title = getString(R.string.weight_loss),
+                title = getString(R.string.female_gender),
                 isSelected = false,
                 type = SignUpSelectorType.RADIO
             )
         )
 
-        adapter.submitList(goalsList)
-        binding.rvGoals.adapter = adapter
+        adapter.submitList(genderList)
+        binding.rvGender.adapter = adapter
+
     }
 }
