+++ app/src/main/java/com/mobven/fitai/login/sign_up/screens/PreferredSportFragment.kt
@@ -1,4 +1,4 @@
-package com.mobven.fitai.login.sign_up
+package com.mobven.fitai.login.sign_up.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,64 +6,57 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.mobven.fitai.R
-import com.mobven.fitai.databinding.FragmentSportBodyPartBinding
+import com.mobven.fitai.databinding.FragmentGoalsBinding
 import com.mobven.fitai.login.sign_up.adapter.SignUpListAdapter
 import com.mobven.fitai.login.sign_up.model.SignUpSelectorItem
 import com.mobven.fitai.util.SignUpSelectorType
 
+class PreferredSportFragment : Fragment() {
 
-class SportBodyPartFragment : Fragment() {
-
-    private lateinit var binding: FragmentSportBodyPartBinding
+    private lateinit var binding: FragmentGoalsBinding
     private val adapter = SignUpListAdapter()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentSportBodyPartBinding.inflate(inflater, container, false)
+    ): View{
+        binding = FragmentGoalsBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val bodyPartList = listOf(
+        val preferredSportList = listOf(
             SignUpSelectorItem(
                 title = getString(R.string.all),
                 isSelected = false,
                 type = SignUpSelectorType.CHECKBOX
             ),
             SignUpSelectorItem(
-                title = getString(R.string.shoulder_arms),
+                title = getString(R.string.yoga),
                 isSelected = false,
                 type = SignUpSelectorType.CHECKBOX
             ),
             SignUpSelectorItem(
-                title = getString(R.string.chest),
+                title = getString(R.string.fitness),
                 isSelected = false,
                 type = SignUpSelectorType.CHECKBOX
             ),
             SignUpSelectorItem(
-                title = getString(R.string.belly_back),
+                title = getString(R.string.running),
                 isSelected = false,
                 type = SignUpSelectorType.CHECKBOX
             ),
             SignUpSelectorItem(
-                title = getString(R.string.hip),
-                isSelected = false,
-                type = SignUpSelectorType.CHECKBOX
-            ),
-            SignUpSelectorItem(
-                title = getString(R.string.leg),
+                title = getString(R.string.walking),
                 isSelected = false,
                 type = SignUpSelectorType.CHECKBOX
             ),
         ).reversed()
 
-        adapter.submitList(bodyPartList)
-        binding.rvBodyPartSport.adapter = adapter
+        adapter.submitList(preferredSportList)
+        binding.rvGoals.adapter = adapter
 
     }
-
 }
