+++ app/src/main/java/com/mobven/fitai/login/sign_up/screens/SignUpFragment.kt
@@ -1,4 +1,4 @@
-package com.mobven.fitai.login.sign_up
+package com.mobven.fitai.login.sign_up.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.mobven.fitai.R
 
-class WeightSignUpFragment : Fragment() {
+class SignUpFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_weight_sign_up, container, false)
+        return inflater.inflate(R.layout.fragment_sign_up, container, false)
     }
 
 }
