+++ app/src/main/res/layout/fragment_average_sleep.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.sign_up.AverageSleepFragment">
+    tools:context=".login.sign_up.screens.AverageSleepFragment">
 
     <TextView
         android:id="@+id/average_sleep_title"
