+++ app/src/main/res/layout/fragment_average_water.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
-    tools:context=".login.sign_up.AverageWaterFragment">
+    tools:context=".login.sign_up.screens.AverageWaterFragment">
 
     <TextView
         android:id="@+id/average_water_title"
