+++ app/src/main/res/layout/fragment_birthday_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.sign_up.BirthdaySignUpFragment">
+    tools:context=".login.sign_up.screens.BirthdaySignUpFragment">
 
     <TextView
         android:id="@+id/birthday_title"
