+++ app/src/main/res/layout/fragment_daily_movement.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.DailyMovementFragment">
+    tools:context=".login.sign_up.screens.DailyMovementFragment">
 
     <TextView
         android:id="@+id/daily_movement_title"
