+++ app/src/main/res/layout/fragment_diet_type.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.sign_up.DietTypeFragment">
+    tools:context=".login.sign_up.screens.DietTypeFragment">
 
     <TextView
         android:id="@+id/diet_type_title"
