+++ app/src/main/res/layout/fragment_food_allergies.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.sign_up.FoodAllergiesFragment">
+    tools:context=".login.sign_up.screens.FoodAllergiesFragment">
 
     <TextView
         android:id="@+id/food_allergies_title"
