+++ app/src/main/res/layout/fragment_other_health_problem.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.sign_up.OtherHealthProblemFragment">
+    tools:context=".login.sign_up.screens.OtherHealthProblemFragment">
 
     <TextView
         android:id="@+id/other_health_problem_title"
