+++ app/src/main/res/layout/fragment_plans.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.PlansFragment">
+    tools:context=".login.sign_up.screens.PlansFragment">
 
     <TextView
         android:id="@+id/plans_title"
