+++ app/src/main/res/layout/fragment_preferred_sport.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.sign_up.PreferredSportFragment">
+    tools:context=".login.sign_up.screens.PreferredSportFragment">
 
     <TextView
         android:id="@+id/preferred_sport_title"
