+++ app/src/main/res/layout/fragment_weight_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.WeightSignUpFragment">
+    tools:context=".login.sign_up.screens.WeightSignUpFragment">
 
     <TextView
         android:id="@+id/weight_title"
